Moves errorMsg to errors.go
